committing div

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -1,7 +1,7 @@
 -- Desc: creates a function SafeDiv that divides (and returns) the first by the second number or returns 0 if the second number is equal to 0.
 
 DELIMITER $$
-CREATE FUNCTION SafeDiv(IN a INT, IN b INT)
+CREATE FUNCTION SafeDiv(IN a FLOAT, IN b FLOAT)
 RETURNS FLOAT
 BEGIN
     IF a = 0 THEN
